+++ Domain/Entities/MaintenanceServiceCost.cs
@@ -15,7 +15,7 @@ namespace Domain.Entities
         }
         [Key]
         public Guid MaintenanceServiceCostId { get; set; }
-        public double ActuralCost { get; set; }
+        public float ActuralCost { get; set; }
         public DateTime DateTime { get; set; }
         public string? Note { get; set; }
         public string Status { get; set; }
