+++ Domain/Entities/MaintenanceSparePartInfo.cs
@@ -19,8 +19,8 @@ namespace Domain.Entities
         public string MaintenanceSparePartInfoName { get; set; }
         public int Quantity { get; set; }
         public int Discount { get; set; }
-        public double ActualCost { get; set; }
-        public double TotalCost { get; set; }
+        public float ActualCost { get; set; }
+        public float TotalCost { get; set; }
         public DateTime CreatedDate { get; set; }
         public string Status { get; set; }
         public string Note { get; set; }
