+++ Domain/Entities/SparePartsItemCost.cs
@@ -17,7 +17,7 @@ namespace Domain.Entities
 
         [Key]
         public Guid SparePartsItemCostId { get; set; }
-        public double ActuralCost { get; set; }
+        public float ActuralCost { get; set; }
         public DateTime DateTime { get; set; }
         public string? Note { get; set; }
         public string Status { get; set; }
