+++ Infrastructure/Common/Request/RequestMaintananceServices/UpdateServies.cs
@@ -11,6 +11,6 @@ namespace Infrastructure.Common.Request.MaintananceServices
         public string ServiceCareName { get; set; }
         public string ServiceCareDescription { get; set; }
         public string ServiceCareType { get; set; }
-        public double OriginalPrice { get; set; }
+        public float OriginalPrice { get; set; }
     }
 }
