+++ Infrastructure/Common/Request/RequestMaintenanceServiceCost/CreateMaintenanceServiceCost.cs
@@ -8,7 +8,7 @@ namespace Infrastructure.Common.Request.RequestMaintenanceServiceCost
 {
     public class CreateMaintenanceServiceCost
     {
-        public double ActuralCost { get; set; }
+        public float ActuralCost { get; set; }
         public string? Note { get; set; }
         public Guid MaintenanceServiceId { get; set; }
     }
