+++ Infrastructure/Common/Request/RequestMaintenanceServiceInfo/CreateMaintenanceServiceInfoHaveItems.cs
@@ -11,7 +11,7 @@ namespace Infrastructure.Common.Request.RequestMaintenanceServiceInfo
         public Guid? MaintenanceServiceCostId { get; set; }
         public string MaintenanceServiceInfoName { get; set; }
         public int Quantity { get; set; }
-        public double ActualCost { get; set; }
+        public float ActualCost { get; set; }
         public string Note { get; set; }
     }
 }
