+++ Infrastructure/Common/Response/ResponseCost/ResponseSparePartsItemCost.cs
@@ -10,7 +10,7 @@ namespace Infrastructure.Common.Response.ResponseCost
     public class ResponseSparePartsItemCost
     {
         public Guid SparePartsItemCostId { get; set; }
-        public double ActuralCost { get; set; }
+        public float ActuralCost { get; set; }
         public DateTime DateTime { get; set; }
         public string? Note { get; set; }
         public string Status { get; set; }
