+++ Infrastructure/Common/Response/ResponseMainInformation/ResponseMaintenanceInformation.cs
@@ -15,7 +15,7 @@ namespace Infrastructure.Common.Response.ResponseMainInformation
         public string InformationMaintenanceName { get; set; }
         public DateTime CreatedDate { get; set; }
         public DateTime FinishedDate { get; set; }
-        public double TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public string Note { get; set; }
         public Guid? BookingId { get; set; }
         public Guid CustomerCareId { get; set; }
